requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
sqlmodel
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
requests